+++ interp/src/flatten/mod.rs
@@ -3,9 +3,7 @@ pub mod primitives;
 mod structures;
 pub(crate) mod text_utils;
 
-use structures::environment::Environment;
-
-use self::structures::environment::Simulator;
+use structures::environment::{Environment, Simulator};
 
 pub fn flat_main(ctx: &calyx_ir::Context) {
     let i_ctx = flat_ir::control::translator::translate(ctx);
